Extract some classes from NameResolutionAndTypeInference.cs
