+++ Source/DafnyCore/Resolver/NameResolutionAndTypeInference/TypeInferenceChecker.cs
@@ -1,0 +1,17 @@
+using static Microsoft.Dafny.ResolutionErrors;
+
+namespace Microsoft.Dafny;
+
+partial class Resolver {
+  public void ReportWarning(ErrorId errorId, IToken t, string msg, params object[] args) {
+    reporter.Warning(MessageSource.Resolver, errorId, t, msg, args);
+  }
+
+  public void ReportError(ErrorId errorId, IToken t, string msg, params object[] args) {
+    reporter.Error(MessageSource.Resolver, errorId, t, msg, args);
+  }
+
+  public void ReportError(ErrorId errorId, INode t, string msg, params object[] args) {
+    reporter.Error(MessageSource.Resolver, errorId, t, msg, args);
+  }
+}
